Fix #476 issue with price input. Please validate before closing.

+++ oc-includes/osclass/classes/actions/ItemActions.php
@@ -453,7 +453,7 @@ class ItemActions
         $flash_error .= ((!osc_validate_max(number_format($aItem['price'], 0, '', ''), 15))
             ? _m('Price too long.')
               . PHP_EOL : '');
-        $flash_error .= (($aItem['price'] !== null && (int)$aItem['price'] < 0)
+        $flash_error .= (($aItem['price'] !== null && (float)$aItem['price'] < 0)
             ? _m('Price must be positive number.') . PHP_EOL : '');
         $flash_error .= ((!osc_validate_text($aItem['countryName'], 3, false))
             ? _m('Country too short.') . PHP_EOL
